Use the new bmp280 driver for kernel 4.9 (the bmp085 doesn't exist any more)

+++ src/veterod/pressurereader.h
@@ -96,7 +96,7 @@ class PressureReader
          * \param[in] pressure the pressure in 1/100 hPa
          * \return the corrected pressure in 1/100 hPa
          */
-        int calculateSeaLevelPressure(int pressure) const;
+        int calculateSeaLevelPressure(double pressure) const;
 
     private:
         std::string m_filename;
